moving infusion functions into base model
