+++ .gitignore
@@ -119,3 +119,4 @@ dmypy.json
 #vscode
 .vscode
 exit
+.pytest_cache
